// tsnode/model/cartData.ts
/**
 * splitit-web-api-public-sdk
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';
import { AmountDetails } from './amountDetails';
import { ItemData } from './itemData';

export class CartData {
    'items'?: Array<ItemData>;
    'amountDetails'?: AmountDetails;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "items",
            "baseName": "Items",
            "type": "Array<ItemData>"
        },
        {
            "name": "amountDetails",
            "baseName": "AmountDetails",
            "type": "AmountDetails"
        }    ];

    static getAttributeTypeMap() {
        return CartData.attributeTypeMap;
    }
}